Made requested changes: Moved unit tests to separate classes and combined Success and Failure listeners with toast.
